backend/requirements.txt
Flask==3.0.3
gunicorn==20.0.4
flask-cors==3.0.10
python-dotenv
flask_migrate
flask_sqlalchemy
sqlalchemy
psycopg2-binary
pyjwt